requirements.txt
Boruta
fastparquet
feature_engine
lightgbm
mlxtend
numpy==1.23.3
pandas==1.5.3
pathlib
pre-commit==2.15.0
pyarrow
pytest==5.4.1
python-dotenv==0.19.1
sklearn-genetic
torch
torch_geometric
xgboost
